+++ backend/src/schemas/reserva.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 class ReservaModel(BaseModel):
     nome: str
     classificacao: int
-    avaliacao: int
+    avaliacao: float
     estado: str
     cidade: str
 
@@ -11,7 +11,7 @@ class ReservaGet(BaseModel):
     id: str
     nome: str
     classificacao: int
-    avaliacao: int
+    avaliacao: float
     estado: str
     cidade: str
